+++ src/main/java/com/kaylves/interfacex/common/resolver/RabbitMQProducerResolver.java
@@ -1,5 +1,6 @@
 package com.kaylves.interfacex.common.resolver;
 
+import com.kaylves.interfacex.annotations.InterfaceXEnum;
 import com.kaylves.interfacex.method.HttpMethod;
 import com.kaylves.interfacex.navigator.RestServiceItem;
 import com.kaylves.interfacex.utils.PsiAnnotationHelper;
@@ -81,7 +82,7 @@ public class RabbitMQProducerResolver extends BaseServiceResolver {
 
         String requestMethod = HttpMethod.PRODUCE.name();
 
-        results.add(new RestServiceItem(method, requestMethod, method.getName(), false));
+        results.add(new RestServiceItem(method, InterfaceXEnum.RabbitMQProducer, requestMethod, method.getName(), false));
 
     }