+++ src/main/java/com/kaylves/interfacex/common/resolver/RocketMQProducerResolver.java
@@ -1,5 +1,6 @@
 package com.kaylves.interfacex.common.resolver;
 
+import com.kaylves.interfacex.annotations.InterfaceXEnum;
 import com.kaylves.interfacex.annotations.rocketmq.RocketMQProducerAnnotation;
 import com.kaylves.interfacex.utils.PsiAnnotationHelper;
 import com.kaylves.interfacex.method.HttpMethod;
@@ -81,7 +82,7 @@ public class RocketMQProducerResolver extends BaseServiceResolver {
 
                 String path = MessageFormat.format("{0}",tags);
 
-                RestServiceItem item = new RestServiceItem(psiMethod, requestMethod, path, false);
+                RestServiceItem item = new RestServiceItem(psiMethod, InterfaceXEnum.RocketMQProducer, requestMethod, path, false);
                 itemList.add(item);
             }
         }