+++ src/main/java/com/kaylves/interfacex/common/resolver/SpringMVCResolver.java
@@ -1,5 +1,6 @@
 package com.kaylves.interfacex.common.resolver;
 
+import com.kaylves.interfacex.annotations.InterfaceXEnum;
 import com.kaylves.interfacex.annotations.http.PathMappingAnnotation;
 import com.kaylves.interfacex.annotations.http.SpringControllerAnnotation;
 import com.kaylves.interfacex.common.spring.RequestMappingAnnotationHelper;
@@ -73,7 +74,7 @@ public class SpringMVCResolver extends BaseServiceResolver {
 
                 for (RequestPath methodRequestPath : methodRequestPaths) {
                     String path = classRequestPath.getPath();
-                    RestServiceItem item = createRestServiceItem(psiMethod, path, methodRequestPath);
+                    RestServiceItem item = createRestServiceItem(psiMethod, InterfaceXEnum.HTTP, path, methodRequestPath);
                     itemList.add(item);
                 }
             }